Handle legacy encryption and silently upgrade
